+++ pybamm/models/submodels/electrolyte/stefan_maxwell/conductivity/full_stefan_maxwell_conductivity.py
@@ -39,7 +39,8 @@ class Full(BaseModel):
         phi_e = variables["Electrolyte potential"]
 
         i_e = (param.kappa_e(c_e, T) * (eps ** param.b) * param.gamma_e / param.C_e) * (
-            param.chi(c_e) * pybamm.grad(c_e) / c_e - pybamm.grad(phi_e)
+            param.chi(c_e) * (1 + param.Theta * T) * pybamm.grad(c_e) / c_e
+            - pybamm.grad(phi_e)
         )
 
         variables.update(self._get_standard_current_variables(i_e))